+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 
 - add convenient template for creating strong types easily
 
+### Type Traits
+
+- Introduce new concept `HasToString`
+
 <!-- insertion marker -->
 ## [0.1.5](https://github.com/repo/owner/releases/tag/0.1.5) - 2026-08-13